server/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
stripe==2.54.0

psycopg2-binary==2.8.5 # работа с базой данных postgres
dj-database-url==0.5.0 # передача настроек БД из env окружения

gunicorn==20.0.4 # http сервер, который свзявает nginx с django приложением